+++ changelog.md
@@ -1,10 +1,14 @@
 # Changelog
-## v0.3.0
+## v0.4
+* Viewer now works Windows 7 using Python 3.8
+* If pygame is not detected it will ask you to install it
+
+## v0.3
 March 5, 2022
 * Sprites now render with transparency
 * Added window title and icon
 
-## v0.2.0
+## v0.2
 March 5, 2022
 * Use the arrow keys to move between maps
 * Press G to toggle the grid overlay
@@ -12,7 +16,7 @@ March 5, 2022
 * JSON parser will skip tiles if they throw an error
 * Fixed some filenames and bad map data in 0106.txt
 
-## v0.1.0
+## v0.1
 March 4, 2022
 * First implementation of map renderer
 * Maps 0106, 0106b, 0306, 0502, and 0506b throw an error when parsing tile JSON